update DESCRIPTION
Note to self: Writing R Extensions says titles of books and the like should be in double
quotes - but CRAN will still complain if double-quoted words are not standard English.
Is OpenStreetMap more like a book, or more like software? Either way, single-quote is the
only way to get CRAN to ignore possibly mis-spelled - that is, non-English - words.